Add Rust native symbolization library and C API wrapper
This commit adds the Rust native symbolization framework that we developed
at Elastic and now making available in OpenTelemetry as part of the
ebpf-profiler project. We can use it to implement automatic on-target
symbolization of native binaries (e.g. Go executables) but also to enable an
OpenTelemetry bulk native symbol upload protocol.

Co-authored-by: Joel Höner <joel@elastic.co>
Co-authored-by: Victor Michel <victor.michel@elastic.co>
Co-authored-by: Florian Lehner <florian.lehner@elastic.co>
Co-authored-by: Tim Rühsen <tim.ruhsen@elastic.co>
Co-authored-by: Davide Girardi <davide.girardi@elastic.co>